pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zhongwen"
version = "5.35.14"
authors = [
  { name="fhopecc", email="fhopecc@gmail.com" },
]
description = "以中文命名之文、數、時、表及檔等處理工具集。"
readme = "README.md"
license = {text = "GPL-3.0-only"} 
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ['aiofiles'
               ,'backupy'
               ,'beautifulsoup4'
               ,'build'
               ,"clipboard"
               ,"comtypes"
               ,"diskcache"
               ,'faker'
               ,'functions-framework'
               ,'furigana' # 日文假名標注
               ,"hanlp" # 中文分詞
               ,'holidays' # 世界各國假日檔
               ,'html5lib'
               ,'jedi' 
               ,'jinja2'
               ,"lark"
               ,'lxml' 
               ,"marisa_trie"
               ,'matplot'
               ,'oauth2client'
               ,'odfpy'
               ,'openai-whisper'  # 中文簡繁轉換
               ,'opencc-python-reimplemented'  # 中文簡繁轉換
               ,'openpyxl'
               ,'orjson'
               ,"pandas"
               ,"pdf2image"
               ,"prophet"
               ,'py7zr'
               ,'pygsheets'
               ,"pygtrans"
               ,'pygtrie'
               ,'pyinstall'
               ,"pypdf"
               ,'pyxirr'
               ,"pytesseract"
               ,"python-docx"
               ,"pywin32"
               ,"PyMuPDF"
               ,"PyPDF2"
               ,"requests"
               ,'scikit-learn'
               ,"selenium"
               ,'sympy'
               ,'tabulate'
               ,'torch'
               ,'torchvision'
               ,'torchaudio'
               ,'treelib'
               ,'twine'
               ,'wcwidth' # 使tabulate 支援中文表格輸出
               ,'webdriver_manager'
               ,"winshell"
               ,'windows_tools'
               ,'winshell'
               ,'wmi'
               ,'xlrd'
               ,'yfinance'
]
[project.urls]
"Homepage" = "https://github.com/fhopecc/zhongwen"
"Bug Tracker" = "https://github.com/fhopecc/zhongwen/issues"
